Implement circuit asserting g1 and g2 point aren't at infinity and pk isn't the neg generator
