+++ beacon-light-client/plonky2/crates/circuits/src/bls_verification/mod.rs
@@ -1,2 +1,3 @@
+pub mod are_points_infinite_circuit;
 pub mod bls12_381_circuit;
 pub mod build_stark_proof_verifier;
